requirements.txt
Django==5.0.6
django-allauth==0.63.3
django-crispy-forms==2.2
python-decouple==3.8
daphne==4.1.2
captcha==0.5.0
django-simple-captcha==0.6.0
channels==4.1.0
requests==2.32.3
PyJWT==2.8.0
pillow==10.4.0
cryptography==42.0.8
django-recaptcha==4.0.0
six==1.16.0
crispy-bootstrap4==2024.1